moving our data structures to classes
