+++ osiris_utils/__init__.py
@@ -1,5 +1,6 @@
 from .utils import (time_estimation, filesize_estimation, transverse_average, integrate, animate_2D,
                     save_data, read_data, mft_decomposition, courant2D)
-from .data import create_dataset, open1D, open2D, open3D, read_osiris_file
+from .data_readers import create_dataset, open1D, open2D, open3D, read_osiris_file
 from .gui import LAVA
 from .analysis import TwoFluid2D
+from .data import OsirisGridFile
